backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

# Redis
redis==5.2.1

# Vector database
qdrant-client==1.12.1

# Google Cloud and AI
google-cloud-storage==2.19.0
google-genai==0.5.0

# HTTP client
httpx==0.28.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Settings
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Development
black==24.10.0
isort==5.13.2
mypy==1.14.1
